refactor(capabilities): migrate the lifecycle cap to the struct-hosted actor macro
Apply the ADR-0123 struct-hosted transform to the lifecycle cap (#2380).
